+++ src/editor/Warp.NET.Editor/Components/Styles/TextButtonStyle.cs
@@ -1,6 +1,0 @@
-using Warp.NET.Numerics;
-using Warp.NET.Text;
-
-namespace Warp.NET.Editor.Components.Styles;
-
-public readonly record struct TextButtonStyle(Color TextColor, TextAlign TextAlign);