+++ src/editor/Warp.NET.Editor/Game.cs
@@ -2,17 +2,17 @@ using Silk.NET.OpenGL;
 using System.Numerics;
 using Warp.NET.Debugging;
 using Warp.NET.Editor.Layouts;
-using Warp.NET.Editor.Rendering;
-using Warp.NET.Editor.Rendering.Renderers;
 using Warp.NET.Extensions;
 using Warp.NET.Numerics;
+using Warp.NET.RenderImpl.Ui;
+using Warp.NET.RenderImpl.Ui.Rendering;
 using Warp.NET.Text;
 using Warp.NET.Ui;
 
 namespace Warp.NET.Editor;
 
 [GenerateGame]
-public sealed partial class Game
+public sealed partial class Game : RenderImplUiGameBase
 {
 	private readonly Matrix4x4 _projectionMatrix;
 
@@ -25,11 +25,6 @@ public sealed partial class Game
 
 		DebugStack.DisplaySetting = DebugStackDisplaySetting.Simple;
 	}
-
-	public MonoSpaceFontRenderer MonoSpaceFontRenderer { get; } = new(new(Textures.Spleen6x12, Charsets.Ascii_32_126));
-	public SpriteRenderer SpriteRenderer { get; } = new();
-	public RectangleRenderer RectangleRenderer { get; } = new();
-	public CircleRenderer CircleRenderer { get; } = new();
 
 	protected override void Update()
 	{
@@ -56,16 +51,16 @@ public sealed partial class Game
 		Gl.ClearColor(0, 0, 0, 1);
 		Gl.Clear(ClearBufferMask.ColorBufferBit);
 
-		Shaders.Ui.Use();
+		RenderImplUiShaders.Ui.Use();
 		Shader.SetMatrix4x4(UiUniforms.Projection, _projectionMatrix);
 		RectangleRenderer.Render();
 		CircleRenderer.Render();
 
-		Shaders.Font.Use();
+		RenderImplUiShaders.Font.Use();
 		Shader.SetMatrix4x4(FontUniforms.Projection, _projectionMatrix);
 		MonoSpaceFontRenderer.Render();
 
-		Shaders.Sprite.Use();
+		RenderImplUiShaders.Sprite.Use();
 		Shader.SetMatrix4x4(SpriteUniforms.Projection, _projectionMatrix);
 		SpriteRenderer.Render();
 	}