+++ src/editor/Warp.NET.Editor/Layouts/MainLayout.cs
@@ -1,6 +1,6 @@
-using Warp.NET.Editor.Components;
-using Warp.NET.Editor.Components.Styles;
-using Warp.NET.Editor.Rendering;
+using Warp.NET.RenderImpl.Ui.Components;
+using Warp.NET.RenderImpl.Ui.Components.Styles;
+using Warp.NET.RenderImpl.Ui.Rendering;
 using Warp.NET.Ui;
 
 namespace Warp.NET.Editor.Layouts;
@@ -8,9 +8,9 @@ namespace Warp.NET.Editor.Layouts;
 public class MainLayout : Layout
 {
 	public MainLayout()
-		: base(Rectangle.Full)
+		: base(Constants.RectangleFull)
 	{
-		Button button = new(new Rectangle(0.1f, 0.1f, 0.1f, 0.1f), () => {}, ButtonStyle.Default);
+		Button button = new(new Rectangle(0.1f, 0.1f, 0.1f, 0.1f, Constants.GridDefault), () => {}, ButtonStyle.Default);
 		NestingContext.Add(button);
 	}
 }