+++ src/lib/Warp.NET.RenderImpl.Ui/Rendering/CoordinateSystem.cs
@@ -1,0 +1,17 @@
+using System.Numerics;
+using Warp.NET.Numerics;
+
+namespace Warp.NET.RenderImpl.Ui.Rendering;
+
+public static class CoordinateSystem
+{
+	public static Vector2i<int> Get(Vector2 position)
+	{
+		return Get(position.X, position.Y);
+	}
+
+	public static Vector2i<int> Get(float x, float y)
+	{
+		return new((int)(x * WarpBase.Game.InitialWindowWidth), (int)(y * WarpBase.Game.InitialWindowHeight));
+	}
+}