+++ src/lib/Warp.NET.RenderImpl.Ui/Rendering/Grid.cs
@@ -1,0 +1,3 @@
+namespace Warp.NET.RenderImpl.Ui.Rendering;
+
+public record Grid(int Width, int Height);